+++ src/FSharpPlus/Functor.fs
@@ -256,32 +256,32 @@ type Map =
         
 
 
-type MZero =
+type Empty =
     inherit Default1
-    static member inline MZero ([<Optional>]_output : '``FunctorZero<'T>``, [<Optional>]_mthd : Default1) = (^``FunctorZero<'T>`` : (static member MZero: ^``FunctorZero<'T>``) ()) : '``FunctorZero<'T>``
-    static member        MZero ([<Optional>]_output : option<'T>          , [<Optional>]_mthd : MZero   ) = None                  : option<'T>
-    static member        MZero ([<Optional>]_output : list<'T>            , [<Optional>]_mthd : MZero   ) = [  ]                  : list<'T>  
-    static member        MZero ([<Optional>]_output : 'T []               , [<Optional>]_mthd : MZero   ) = [||]                  : 'T []     
-    static member        MZero ([<Optional>]_output : seq<'T>             , [<Optional>]_mthd : MZero   ) = Seq.empty             : seq<'T>
-    static member inline MZero ([<Optional>]_output : Id<'T>              , [<Optional>]_mthd : MZero   ) = Id (MEmpty.Invoke())   : Id<'T>
+    static member inline Empty ([<Optional>]_output : '``FunctorZero<'T>``, [<Optional>]_mthd : Default1) = (^``FunctorZero<'T>`` : (static member Empty: ^``FunctorZero<'T>``) ()) : '``FunctorZero<'T>``
+    static member        Empty ([<Optional>]_output : option<'T>          , [<Optional>]_mthd : Empty   ) = None                  : option<'T>
+    static member        Empty ([<Optional>]_output : list<'T>            , [<Optional>]_mthd : Empty   ) = [  ]                  : list<'T>  
+    static member        Empty ([<Optional>]_output : 'T []               , [<Optional>]_mthd : Empty   ) = [||]                  : 'T []     
+    static member        Empty ([<Optional>]_output : seq<'T>             , [<Optional>]_mthd : Empty   ) = Seq.empty             : seq<'T>
+    static member inline Empty ([<Optional>]_output : Id<'T>              , [<Optional>]_mthd : Empty   ) = Id (MEmpty.Invoke())  : Id<'T>
 
     static member inline Invoke () : '``FunctorZero<'T>`` =
-        let inline call (mthd : ^M, output : ^R) = ((^M or ^R) : (static member MZero: _*_ -> _) output, mthd)
-        call (Unchecked.defaultof<MZero>, Unchecked.defaultof<'``FunctorZero<'T>``> )
+        let inline call (mthd : ^M, output : ^R) = ((^M or ^R) : (static member Empty: _*_ -> _) output, mthd)
+        call (Unchecked.defaultof<Empty>, Unchecked.defaultof<'``FunctorZero<'T>``> )
 
 
-type MPlus =
+type Append =
     inherit Default1
-    static member inline MPlus (x :'``FunctorPlus<'T>``, y:'``FunctorPlus<'T>``, [<Optional>]_mthd : Default1) = (^``FunctorPlus<'T>`` :  (static member MPlus : _*_ -> _) x, y) : ^``FunctorPlus<'T>``
-    static member        MPlus (x :'T option           , y                     , [<Optional>]_mthd : MPlus   ) = match x with None -> y | xs -> xs
-    static member        MPlus (x :'T list             , y                     , [<Optional>]_mthd : MPlus   ) = x @ y
-    static member        MPlus (x :'T []               , y                     , [<Optional>]_mthd : MPlus   ) = Array.append x y
-    static member        MPlus (x :'T seq              , y                     , [<Optional>]_mthd : MPlus   ) = Seq.append   x y
-    static member inline MPlus (x :'T Id               , y                     , [<Optional>]_mthd : MPlus   ) = Id (MAppend.Invoke (Id.run x) (Id.run y))
+    static member inline Append (x :'``FunctorPlus<'T>``, y:'``FunctorPlus<'T>``, [<Optional>]_mthd : Default1) = (^``FunctorPlus<'T>`` :  (static member Append : _*_ -> _) x, y) : ^``FunctorPlus<'T>``
+    static member        Append (x :'T option           , y                     , [<Optional>]_mthd : Append  ) = match x with None -> y | xs -> xs
+    static member        Append (x :'T list             , y                     , [<Optional>]_mthd : Append  ) = x @ y
+    static member        Append (x :'T []               , y                     , [<Optional>]_mthd : Append  ) = Array.append x y
+    static member        Append (x :'T seq              , y                     , [<Optional>]_mthd : Append  ) = Seq.append   x y
+    static member inline Append (x :'T Id               , y                     , [<Optional>]_mthd : Append  ) = Id (MAppend.Invoke (Id.run x) (Id.run y))
 
     static member inline Invoke (x:'``FunctorPlus<'T>``) (y:'``FunctorPlus<'T>``)  : '``FunctorPlus<'T>`` =
-        let inline call (mthd : ^M, input1 : ^I, input2 : ^I) = ((^M or ^I) : (static member MPlus: _*_*_ -> _) input1, input2, mthd)
-        call (Unchecked.defaultof<MPlus>, x, y)
+        let inline call (mthd : ^M, input1 : ^I, input2 : ^I) = ((^M or ^I) : (static member Append: _*_*_ -> _) input1, input2, mthd)
+        call (Unchecked.defaultof<Append>, x, y)
 
 
 type Delay =
@@ -807,5 +807,5 @@ module internal MonadOps =
     let inline (>>=) x f = FsControl.Bind.Invoke x f
     let inline result  x = FsControl.Return.Invoke x
     let inline (<*>) f x = FsControl.Apply.Invoke f x
-    let inline (<|>) x y = FsControl.MPlus.Invoke x y
+    let inline (<|>) x y = FsControl.Append.Invoke x y
     let inline (>=>) (f:'a->'Monad'b) (g:'b->'Monad'c) (x:'a) :'Monad'c = f x >>= g