+++ src/FSharpPlus/Monoids.fs
@@ -85,8 +85,8 @@ type Compose<'``f<'g<'t>>``> = Compose of '``f<'g<'t>>`` with
     static member inline (<*>)  (Compose (f: '``F<'G<'T->'U>>``), Compose (x: '``F<'G<'T>>``)) = Compose ((<*>) <!> f <*> x: '``F<'G<'U>>``)
 
     // Alternative
-    static member inline get_MZero()                  = Compose (getMZero()) : Compose<'``F<'G<'T>``>
-    static member inline MPlus (Compose x, Compose y) = Compose (x <|> y)    : Compose<'``F<'G<'T>``>
+    static member inline get_Empty()                  = Compose (getEmpty()) : Compose<'``F<'G<'T>``>
+    static member inline Append (Compose x, Compose y) = Compose (x <|> y)    : Compose<'``F<'G<'T>``>
 
 
 /// Basic operations on Compose