+++ src/FSharpPlus/Option.fs
@@ -26,8 +26,8 @@ type OptionT with
     static member inline (<*>)  (f : OptionT<'``Monad<seq<('T -> 'U)>``>, x : OptionT<'``Monad<seq<'T>``>) = OptionT.apply f x  : OptionT<'``Monad<seq<'U>``>
     static member inline Bind   (x : OptionT<'``Monad<seq<'T>``>, f : 'T -> OptionT<'``Monad<seq<'U>``>)   = OptionT.bind f x
 
-    static member inline get_MZero () = OptionT <| result None : OptionT<'``MonadPlus<option<'T>``>
-    static member inline MPlus (OptionT x, OptionT y) = OptionT <| (x  >>= (fun maybe_value -> match maybe_value with Some _ -> x | _ -> y)) : OptionT<'``MonadPlus<option<'T>``>
+    static member inline get_Empty () = OptionT <| result None : OptionT<'``MonadPlus<option<'T>``>
+    static member inline Append (OptionT x, OptionT y) = OptionT <| (x  >>= (fun maybe_value -> match maybe_value with Some _ -> x | _ -> y)) : OptionT<'``MonadPlus<option<'T>``>
 
     static member inline Lift (x:'``Monad<'T>``) = x |> liftM Some |> OptionT : OptionT<'``Monad<option<'T>>``>